Fix build warnings

+++ src/compiler/parse_table.h
@@ -13,7 +13,7 @@
 
 namespace tree_sitter {
 
-typedef int64_t ParseStateId;
+typedef size_t ParseStateId;
 
 enum ParseActionType {
   ParseActionTypeError,
